fix(M5): GET /usage の SUM(bigint)→NUMERIC デコード不一致を修正
live E2E（docker compose 上で実際に invoke → GET /usage）で発見。Postgres は
SUM(bigint) を NUMERIC で返すため、集計読み取りクエリの結果を Rust の i64
（UsageRollupRow）にデコードしようとして 500（"mismatched types ... i64 (INT8)
is not compatible with NUMERIC"）になっていた。rollup が空のときは行が無く露見せず、
1 件でも計上されると必ず失敗する状態だった。

各 SUM(...) を ::bigint へ明示キャストして i64 デコードと型を一致させる（MAX は
元の bigint を保つ）。saturating_i64 で書込時に頭打ち済みのため実運用域で i64 に収まる。

DB-free 単体テストでは検出不能なクラスのバグ（要 live DB）。chaos_m5 の E2
（rollup 挿入後に GET /usage を読む）が今後の回帰ガードになる。

検証: docker compose スタックで bootstrap→deploy→invoke→GET /usage が
計量を正しく返し（wall_time_ms/peak_memory_bytes_max/output_bytes 実測）、
chaos_m5 の E1（重複は +1）/ E2（N 件は +N）が live で 2/2 pass。

+++ crates/control-plane/src/db.rs
@@ -962,15 +962,18 @@ pub async fn get_usage_rollups(
     component_id: Option<&str>,
 ) -> Result<Vec<UsageRollupRow>, sqlx::Error> {
     let rows = sqlx::query(
+        // SUM(bigint) は Postgres では NUMERIC を返すため、各集計を ::bigint へ明示キャストして
+        // Rust 側の i64 デコード（UsageRollupRow）と型を一致させる（MAX は元の bigint を保つ）。
+        // 行数 × 各列とも実運用域では i64 に収まる（saturating_i64 で書込時に頭打ち済み）。
         "SELECT component_id, \
-                SUM(invocation_count) AS invocation_count, \
-                SUM(cpu_fuel_used) AS cpu_fuel_used, \
-                SUM(wall_time_ms) AS wall_time_ms, \
+                SUM(invocation_count)::bigint AS invocation_count, \
+                SUM(cpu_fuel_used)::bigint AS cpu_fuel_used, \
+                SUM(wall_time_ms)::bigint AS wall_time_ms, \
                 MAX(peak_memory_bytes_max) AS peak_memory_bytes_max, \
-                SUM(output_bytes) AS output_bytes, \
-                SUM(succeeded_count) AS succeeded_count, \
-                SUM(failed_count) AS failed_count, \
-                SUM(timeout_count) AS timeout_count \
+                SUM(output_bytes)::bigint AS output_bytes, \
+                SUM(succeeded_count)::bigint AS succeeded_count, \
+                SUM(failed_count)::bigint AS failed_count, \
+                SUM(timeout_count)::bigint AS timeout_count \
          FROM usage_rollups \
          WHERE tenant_id = $1 AND period_start >= $2 AND period_start <= $3 \
            AND ($4::text IS NULL OR component_id = $4) \
